Added Basic Thrust and Rotation to Rocket
Added thrust to the rocket in Rocket Boost Turbo. Did this by first getting a reference to the rigidbody component of the rocket, and then using the addRelativeForce method of rigidbodies to give it thrust when the space bar is pressed. Also added in rotation to the transform of the rocket, allowing it to twist and wiggle.