Refactored Code and fixed movement bugs
Got rigidbody movement bugs fixed with rigidbody.freezeRotation property. Otherwise whenever you hit an object the rigidbody continues to be affected, since we are manipulating the rocket through only the transform. Also made some tweaks to the rocket mass and changed the look of the ship slightly. Finally, refactored the code to be cleaner and more concise.