Added first Tags. Fixed Rigidbody Constraint bug
Added in tags to the game that print to the console when you collide with something. Also added in a trigger testing when passing through a fuel representing object. Finally, modified the code to fix a bug that was causing the rocket to not be constrained in the 2d Plane (See lines 84-87 and 103-105